10/28 generator developed 0.0.1
이슈 해결 완료 간단하게 고칠 수 있었다.
서비스 계층에서 트랜잭션을 나누어 실행시키면 되는 간단한 문제였다. 식별 엔티티 간의 키매핑 때문에 수정된 부분이 있다.

+++ genenrator/src/main/java/picto/com/generator/domain/photo/domain/Photo.java
@@ -34,10 +34,10 @@ public class Photo {
     private String location;
 
     @Column(name = "register_time", nullable = false)
-    private int registerTime;
+    private Long registerTime;
 
     @Column(name = "upload_time", nullable = false)
-    private int uploadTime;
+    private Long uploadTime;
 
     // 액자 활성화 여부
     @ColumnDefault("true")
